# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Physics Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses/{slug}:
    get:
      operationId: getCourse
      tags: [courses]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Course with sections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseWithSections"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sections/{sectionId}/topics:
    get:
      operationId: listTopics
      tags: [courses]
      parameters:
        - name: sectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Topics in section
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"
  /topics/{topicId}:
    get:
      operationId: getTopic
      tags: [courses]
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Topic detail with elements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: No access to course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── PROGRESS ────────────────────────────────────────────────────────────────
...
  /ai/check:
    post:
      operationId: checkTask
      tags: [ai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCheckInput"
      responses:
        "200":
          description: AI feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCheckResult"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai/lesson-chat:
    post:
      operationId: lessonChat
      tags: [ai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonChatInput"
      responses:
        "200":
          description: AI assistant reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonChatResult"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── PAYMENTS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Section:
      type: object
      required: [id, courseId, title, slug, sortOrder, createdAt]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        slug:
          type: string
        sortOrder:
          type: integer
        topicCount:
          type: integer
        createdAt:
          type: string
    CourseWithSections:
      type: object
      required: [id, title, slug, description, isPublished, sections]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        isPublished:
          type: boolean
        sections:
          type: array
          items:
            $ref: "#/components/schemas/Section"
    Topic:
      type: object
      required: [id, sectionId, title, slug, sortOrder, createdAt]
      properties:
        id:
          type: integer
        sectionId:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isPreview:
          type: boolean
        hasVideo:
          type: boolean
        hasQuiz:
          type: boolean
        hasTasks:
          type: boolean
        createdAt:
          type: string
    Video:
      type: object
      required: [id, topicId, title, createdAt]
      properties:
        id:
          type: integer
        topicId:
          type: integer
        bunnyVideoId:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        embedUrl:
          type: ["string", "null"]
        title:
          type: string
        durationSeconds:
          type: ["integer", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
    LessonImage:
      type: object
      required: [id, topicId, imageUrl, sortOrder]
      properties:
        id:
          type: integer
        topicId:
          type: integer
        imageUrl:
          type: string
        alt:
          type: ["string", "null"]
        sortOrder:
          type: integer
    Task:
      type: object
      required: [id, topicId, title, createdAt]
      properties:
        id:
          type: integer
        topicId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        initialImageUrl:
          type: ["string", "null"]
        aiPromptConfig:
          type: ["object", "null"]
        createdAt:
          type: string
    TopicDetail:
      type: object
      required: [id, sectionId, title, slug, sortOrder, videos, images, tasks]
      properties:
        id:
          type: integer
        sectionId:
          type: integer
        courseId:
          type: ["integer", "null"]
        title:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isPreview:
          type: boolean
        previousTopicId:
          type: ["integer", "null"]
        nextTopicId:
          type: ["integer", "null"]
        video:
          oneOf:
            - $ref: "#/components/schemas/Video"
            - type: "null"
        videos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
        images:
          type: array
          items:
            $ref: "#/components/schemas/LessonImage"
        quiz:
          oneOf:
            - $ref: "#/components/schemas/QuizPublic"
            - type: "null"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
...
    QuizAnswerPublic:
      type: object
      required: [id, questionId, answerLabel, answerText]
      properties:
        id:
          type: integer
        questionId:
          type: integer
        answerLabel:
          type: string
        answerText:
          type: string
    QuizQuestionPublic:
      type: object
      required: [id, quizId, questionText, sortOrder, answers]
      properties:
        id:
          type: integer
        quizId:
          type: integer
        questionText:
          type: string
        sortOrder:
          type: integer
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswerPublic"
    QuizPublic:
      type: object
      required: [id, topicId, title, questions]
      properties:
        id:
          type: integer
        topicId:
          type: integer
        title:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestionPublic"
...
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
    LessonChatInput:
      type: object
      required: [topicId, message]
      properties:
        topicId:
          type: integer
        message:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
    LessonChatResult:
      type: object
      required: [reply]
      properties:
        reply:
          type: string
...
    TaskCheckInput:
      type: object
      required: [taskId, imageBase64]
      properties:
        taskId:
          type: integer
        imageBase64:
          type: string
    TaskCheckResult:
      type: object
      required: [feedback]
      properties:
        feedback:
          type: string
        checkId:
          type: ["integer", "null"]

    # ─── PAYMENT SCHEMAS ──────────────────────────────────────────────────────
...